feat: add language constant extern file
