+++ src/components/UserLogin/UserLogin.tsx
@@ -3,21 +3,22 @@ import React from 'react';
 import { Button } from '@mui/material';
 import { getAuth, signOut } from 'firebase/auth';
 import { useAppSelector } from '../../hook';
+import text from '../../language/Language';
 
-interface TextKey {
-  logout: string;
-}
-interface Text {
-  [key: string]: TextKey;
-}
-const text: Text = {
-  en: {
-    logout: 'log out',
-  },
-  ru: {
-    logout: 'выйти',
-  },
-};
+// interface TextKey {
+//   logout: string;
+// }
+// interface Text {
+//   [key: string]: TextKey;
+// }
+// const text: Text = {
+//   en: {
+//     logout: 'log out',
+//   },
+//   ru: {
+//     logout: 'выйти',
+//   },
+// };
 
 function LoginRowBtn() {
   const lang = useAppSelector((state) => state.langState.lang);
